# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/instagram/hashtags": {
            "get": {
                "description": "Searches for Instagram hashtags based on a query string",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Instagram"
                ],
                "summary": "Search Instagram hashtags",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search query for hashtags",
                        "name": "query",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/instagram/timeline": {
            "get": {
                "description": "Fetches timeline/posts for a given Instagram username",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Instagram"
                ],
                "summary": "Get Instagram user timeline",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Instagram username",
                        "name": "username",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/instagram/userinfo": {
            "get": {
                "description": "Fetches detailed user information for a given Instagram username",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Instagram"
                ],
                "summary": "Get Instagram user information",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Instagram username (default: sakshi_raghu_1c_)",
                        "name": "username",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/instagram.UserInfoResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/snapchat/curatedhighlights": {
            "get": {
                "description": "Fetches curated highlights for a given Snapchat username",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Snapchat"
                ],
                "summary": "Get Snapchat user curated highlights",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Snapchat username (default: arora_girl)",
                        "name": "username",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/snapchat.CuratedHighlightsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/snapchat/spotlighthighlights": {
            "get": {
                "description": "Fetches spotlight highlights for a given Snapchat username",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Snapchat"
                ],
                "summary": "Get Snapchat user spotlight highlights",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Snapchat username (default : arora_girl)",
                        "name": "username",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/snapchat.SpotlightHighlight"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/snapchat/userinfo": {
            "get": {
                "description": "Fetches detailed user information for a given Snapchat username",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Snapchat"
                ],
                "summary": "Get Snapchat user information",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Snapchat username (default: arora_girl)",
                        "name": "username",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/snapchat.UserInfoResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "instagram.BioLink": {
            "type": "object",
            "properties": {
                "link_type": {
                    "type": "string"
                },
                "lynx_url": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "instagram.BiographyWithEntities": {
            "type": "object",
            "properties": {
                "entities": {
                    "description": "Can be parsed into specific entity types if needed",
                    "type": "array",
                    "items": {}
                },
                "raw_text": {
                    "type": "string"
                }
            }
        },
        "instagram.EdgeConnection": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "edges": {
                    "type": "array",
                    "items": {}
                },
                "page_info": {
                    "$ref": "#/definitions/instagram.PageInfo"
                }
            }
        },
        "instagram.EdgeCount": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                }
            }
        },
        "instagram.EdgeMutualFollowedBy": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "edges": {
                    "type": "array",
                    "items": {}
                }
            }
        },
        "instagram.EdgeRelatedProfiles": {
            "type": "object",
            "properties": {
                "edges": {
                    "type": "array",
                    "items": {}
                }
            }
        },
...
        "instagram.UserInfo": {
            "type": "object",
            "properties": {
                "ai_agent_owner_username": {
                    "type": "string"
                },
                "ai_agent_type": {
                    "type": "string"
                },
                "bio_links": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/instagram.BioLink"
                    }
                },
                "biography": {
                    "type": "string"
                },
                "biography_with_entities": {
                    "$ref": "#/definitions/instagram.BiographyWithEntities"
                },
                "blocked_by_viewer": {
                    "type": "boolean"
                },
                "business_address_json": {
                    "type": "string"
                },
                "business_category_name": {
                    "type": "string"
                },
                "business_contact_method": {
                    "type": "string"
                },
                "business_email": {
                    "type": "string"
                },
                "business_phone_number": {
                    "type": "string"
                },
                "category_enum": {
                    "type": "string"
                },
                "category_name": {
                    "type": "string"
                },
                "country_block": {
                    "type": "boolean"
                },
                "edge_follow": {
                    "description": "EdgeFelixVideoTimeline     EdgeConnection        `json:\"edge_felix_video_timeline\"`",
                    "allOf": [
                        {
                            "$ref": "#/definitions/instagram.EdgeCount"
                        }
                    ]
                },
                "edge_followed_by": {
                    "$ref": "#/definitions/instagram.EdgeCount"
                },
                "edge_media_collections": {
                    "$ref": "#/definitions/instagram.EdgeConnection"
                },
                "edge_mutual_followed_by": {
                    "$ref": "#/definitions/instagram.EdgeMutualFollowedBy"
                },
                "edge_related_profiles": {
                    "description": "EdgeOwnerToTimelineMedia   EdgeConnection        `json:\"edge_owner_to_timeline_media\"`",
                    "allOf": [
                        {
                            "$ref": "#/definitions/instagram.EdgeRelatedProfiles"
                        }
                    ]
                },
                "edge_saved_media": {
                    "$ref": "#/definitions/instagram.EdgeConnection"
                },
                "fb_page_calls_count": {
                    "type": "integer"
                },
                "fb_page_email": {
                    "type": "string"
                },
                "followed_by_viewer": {
                    "type": "boolean"
                },
                "follows_viewer": {
                    "type": "boolean"
                },
                "full_name": {
                    "type": "string"
                },
                "group_metadata": {
                    "type": "string"
                },
                "has_adjusted_ads": {
                    "type": "boolean"
                },
                "has_ar_effects": {
                    "type": "boolean"
                },
                "has_blocked_viewer": {
                    "type": "boolean"
                },
                "has_channel": {
                    "type": "boolean"
                },
                "has_guides": {
                    "type": "boolean"
                },
                "has_requested_viewer": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "ig_custom_customer_sales_channel": {
                    "type": "boolean"
                },
                "instafilms_count": {
                    "type": "integer"
                },
                "instagram_location": {
                    "$ref": "#/definitions/instagram.InstagramLocation"
                },
                "is_private": {
                    "type": "boolean"
                },
                "is_profile_private": {
                    "type": "boolean"
                },
                "is_verified": {
                    "type": "boolean"
                },
                "is_verified_by_mv2": {
                    "type": "boolean"
                },
                "joined_recently": {
                    "type": "boolean"
                },
                "likely_has_audience": {
                    "type": "boolean"
                },
                "media_count": {
                    "type": "integer"
                },
                "mutual_followers_count": {
                    "type": "integer"
                },
                "number_of_links_in_bio": {
                    "type": "integer"
                },
                "page_id_for_ads": {
                    "type": "string"
                },
                "page_number": {
                    "type": "integer"
                },
                "profile_context": {
                    "type": "string"
                },
                "profile_pic_id": {
                    "type": "string"
                },
                "profile_pic_url": {
                    "type": "string"
                },
                "profile_pic_url_hd": {
                    "type": "string"
                },
                "profile_viewer_follow_button": {
                    "type": "boolean"
                },
                "promo_share_user_follow_button": {
                    "type": "boolean"
                },
                "requested_by_viewer": {
                    "type": "boolean"
                },
                "restricted_by_viewer": {
                    "type": "boolean"
                },
                "should_show_category": {
                    "type": "boolean"
                },
                "should_show_public_contacts": {
                    "type": "boolean"
                },
                "social_context": {
                    "type": "string"
                },
                "social_context_list": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "status_emojis": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "type": "string"
                },
                "total_clip_count": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                },
                "verification_data": {
                    "$ref": "#/definitions/instagram.VerificationData"
                }
            }
        },
        "instagram.UserInfoResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "properties": {
                        "user": {
                            "$ref": "#/definitions/instagram.UserInfo"
                        }
                    }
                }
            }
        },
...
        "snapchat.BoolValue": {
            "type": "object",
            "properties": {
                "value": {
                    "type": "boolean"
                }
            }
        },
        "snapchat.CuratedHighlight": {
            "type": "object",
            "properties": {
                "canonicalUrlSuffix": {
                    "type": "string"
                },
                "emoji": {
                    "type": "string"
                },
                "highlightId": {
                    "$ref": "#/definitions/snapchat.StringValue"
                },
                "isAttributed": {
                    "$ref": "#/definitions/snapchat.BoolValue"
                },
                "snapList": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/snapchat.CuratedSnap"
                    }
                },
                "storySubtitle": {
                    "type": "string"
                },
                "storyTapId": {
                    "type": "string"
                },
                "storyTitle": {
                    "type": "string"
                },
                "storyType": {
                    "type": "integer"
                },
                "thumbnailUrl": {
                    "$ref": "#/definitions/snapchat.StringValue"
                },
                "videoTrackUrl": {
                    "type": "string"
                }
            }
        },
        "snapchat.CuratedHighlightsResponse": {
            "type": "object",
            "properties": {
                "highlights": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/snapchat.CuratedHighlight"
                    }
                }
            }
        },
...
        "snapchat.SpotlightHighlight": {
            "type": "object",
            "properties": {
                "canonicalUrlSuffix": {
                    "type": "string"
                },
                "emoji": {
                    "type": "string"
                },
                "highlightId": {
                    "type": "string"
                },
                "isAttributed": {
                    "$ref": "#/definitions/snapchat.BoolValue"
                },
                "snapList": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/snapchat.SpotlightSnap"
                    }
                },
                "storyId": {
                    "$ref": "#/definitions/snapchat.StringValue"
                },
                "storyShareId": {
                    "type": "string"
                },
                "storySubtitle": {
                    "type": "string"
                },
                "storyTapId": {
                    "type": "string"
                },
                "storyTitle": {
                    "type": "string"
                },
                "storyType": {
                    "type": "integer"
                },
                "thumbnailUrl": {
                    "$ref": "#/definitions/snapchat.StringValue"
                },
                "videoTrackUrl": {
                    "type": "string"
                }
            }
        },
        "snapchat.SpotlightSnap": {
            "type": "object",
            "properties": {
                "audioTranscriptionObjectUrl": {
                    "type": "string"
                },
                "hasAttachment": {
                    "type": "boolean"
                },
                "intervalStartTimeMs": {
                    "type": "integer"
                },
                "isSponsored": {
                    "type": "boolean"
                },
                "lat": {
                    "type": "number"
                },
                "lng": {
                    "type": "number"
                },
                "snapId": {
                    "$ref": "#/definitions/snapchat.StringValue"
                },
                "snapIndex": {
                    "type": "integer"
                },
                "snapMediaType": {
                    "type": "integer"
                },
                "snapSubtitles": {
                    "type": "string"
                },
                "snapTitle": {
                    "type": "string"
                },
                "snapUrls": {
                    "$ref": "#/definitions/snapchat.SnapURLs"
                },
                "timestampInSec": {
                    "$ref": "#/definitions/snapchat.StringValue"
                }
            }
        },
        "snapchat.StringValue": {
            "type": "object",
            "properties": {
                "value": {
                    "type": "string"
                }
            }
        },
...
        "snapchat.UserInfo": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "badge": {
                    "type": "integer"
                },
                "bio": {
                    "type": "string"
                },
                "bitmoji3d": {
                    "description": "always null in sample"
                },
                "businessProfileId": {
                    "type": "string"
                },
                "categoryStringId": {
                    "type": "string"
                },
                "creationTimestampMs": {
                    "$ref": "#/definitions/snapchat.TimestampWrapper"
                },
                "hasCuratedHighlights": {
                    "type": "boolean"
                },
                "hasSpotlightHighlights": {
                    "type": "boolean"
                },
                "hasStory": {
                    "type": "boolean"
                },
                "lastUpdateTimestampMs": {
                    "$ref": "#/definitions/snapchat.TimestampWrapper"
                },
                "mutableName": {
                    "type": "string"
                },
                "primaryColor": {
                    "type": "string"
                },
                "profilePictureUrl": {
                    "type": "string"
                },
                "publisherType": {
                    "type": "string"
                },
                "relatedAccountsInfo": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/snapchat.RelatedAccountInfo"
                    }
                },
                "sameAsLinks": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "shouldHideUsername": {
                    "type": "boolean"
                },
                "snapcodeImageUrl": {
                    "type": "string"
                },
                "squareHeroImageUrl": {
                    "type": "string"
                },
                "subcategoryStringId": {
                    "type": "string"
                },
                "subscriberCount": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                },
                "websiteUrl": {
                    "type": "string"
                }
            }
        },
        "snapchat.UserInfoResponse": {
            "type": "object",
            "properties": {
                "$case": {
                    "type": "string"
                },
                "publicProfileInfo": {
                    "$ref": "#/definitions/snapchat.UserInfo"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        }
    }
}